Moved ReadOnlyJsonConverter to Base folder
