+++ K2Bridge/JsonConverters/Base/ReadOnlyJsonConverter.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 // See LICENSE file in the project root for full license information.
 
-namespace K2Bridge.JsonConverters
+namespace K2Bridge.JsonConverters.Base
 {
     using System;
     using Newtonsoft.Json;
