+++ K2Bridge/JsonConverters/SortClauseConverter.cs
@@ -5,6 +5,7 @@
 namespace K2Bridge.JsonConverters
 {
     using System;
+    using K2Bridge.JsonConverters.Base;
     using K2Bridge.Models.Request;
     using Newtonsoft.Json;
     using Newtonsoft.Json.Linq;
